+++ src/modules/voucher/handlers/redeemVoucher.handler.ts
@@ -8,6 +8,7 @@ import { PrismaService } from 'prisma/prisma.service';
 import { BlockchainService } from 'src/providers/blockchain/blockchain.service';
 import { TransactionTypeId } from 'src/constants/transaction-types.enum';
 import { AssetType } from '@prisma/client';
+import { randomUUID } from 'crypto';
 import { TokenService } from 'src/providers/token/token.service';
 import { ConfigService } from '@nestjs/config';
 
@@ -310,6 +311,7 @@ export class RedeemVoucher {
             voucherCodeId: voucherCode.id,
             transactionTypeId: TransactionTypeId.REDEEM,
             type: AssetType.VOUCHER,
+            transactionRefId: randomUUID(),
           },
         }),
       ]);