+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## Version History
 
+- **[Unreleased]** - Transaction classification & seller marketplace listings ([#87](https://github.com/corp-ais/blockchain-backoffice-backend/pull/87))
 - **[Unreleased]** - Type interfaces, refactor handlers & new endpoints ([#80](https://github.com/corp-ais/blockchain-backoffice-backend/pull/80))
 - **[Unreleased]** - Test suite improvements & bug fixes ([#78](https://github.com/corp-ais/blockchain-backoffice-backend/pull/78))
 - **[Unreleased]** - Global transaction endpoints + Response refactoring ([#77](https://github.com/corp-ais/blockchain-backoffice-backend/pull/77))